+++ android/src/main/kotlin/com/maru/twitter_login/TwitterLoginPlugin.kt
@@ -2,6 +2,8 @@ package com.maru.twitter_login
 
 
 import android.content.Intent
+import com.maru.twitter_login.chrome_custom_tab.ChromeCustomTabManager
+import com.maru.twitter_login.chrome_custom_tab.CustomTabActivityHelper
 import io.flutter.embedding.android.FlutterActivity
 import io.flutter.embedding.engine.FlutterEngine
