+++ android/src/main/kotlin/com/maru/twitter_login/chrome_custom_tab/CustomTabActivity.java
@@ -1,4 +1,4 @@
-package com.maru.twitter_login;
+package com.maru.twitter_login.chrome_custom_tab;
 
 import android.app.Activity;
 import android.os.Bundle;
